fix(ce-prototype): authorize handoff and preserve feedback delivery
Follow up upstream #1623 and address EveryInc/compound-engineering-plugin#1640.
Port only the prototype changes from fork PR #4 onto the #1623 merge
baseline, excluding the unrelated skill-eval work in upstream #1638.

Separate public viewing from private annotation bootstrap, preserve the
selected same-origin destination, drain piped wait output, and retain the
server cases in independently runnable test files. Require compatible
annotation-protocol metadata when reusing a live server after an in-place
upgrade, with five new upgrade/reuse regressions. Keep both helper copies
byte-identical.

Local Linux Node v22.16.0: exact source blob verified; syntax check passes;
new regression logic fails on the previous reuse path and passes all five
cases on this implementation under node:test. Bun and browser end-to-end
validation are not claimed by these local checks.
