+++ skills/ce-prototype/references/preview.md
@@ -84,7 +84,7 @@ node "$SKILL_DIR/scripts/light-webserver.js" status --root "$PROTO_DIR"
 
 If `SKILL_DIR` cannot be resolved to a concrete skill directory, do not guess from the project CWD. Stop and report that the preview cannot start; do not settle the question in chat instead.
 
-The helper creates `screens/` and `state/`, serves the newest `.html` file in `screens/` at `/`, writes `state/display-info.json`, and exposes `/version` so a default (annotate-off) browser can poll for screen changes. Isolated web starts pass `--annotate`: the printed URL is the origin (no token in it), the overlay is added at serve time to every HTML document the browser navigates to under `screens/` (a script fetching an HTML partial gets it raw), and disk screens stay agent-clean. Visiting that origin sets a session cookie so the overlay can reach wait, annotation, and events; those routes stay token-gated. Do not print the token to the explorer. A prototype may use `?token=` or other query state of its own. Every other path is read from `screens/` at that same path — `/img/blot.webp` serves `screens/img/blot.webp` — so a screen keeps whatever asset layout it was copied from, nesting included. Put the assets the screen references under `screens/` at the paths it asks for, or inline them as data URIs. Anything resolving outside `screens/` is refused.
+The helper creates `screens/` and `state/`, serves the newest `.html` file in `screens/` at `/`, writes `state/display-info.json`, and exposes `/version` so a default (annotate-off) browser can poll for screen changes. Isolated web starts pass `--annotate`: the printed `url` remains the token-free origin, and `authorize_url` is the private first-visit link. That link validates the session token and issues an HttpOnly, SameSite=Strict cookie; a helper-owned page then navigates to the selected screen (`/` by default), without the bootstrap credential, before any authored screen script runs. Only authenticated HTML navigations receive the overlay. Public visits and script fetches get raw screens, without a cookie or overlay, and disk screens stay agent-clean. Wait, annotation, events, flush, and session-end routes remain credential-gated. A prototype may use `?token=` or other query state of its own. Every other path is read from `screens/` at that same path — `/img/blot.webp` serves `screens/img/blot.webp` — so a screen keeps whatever asset layout it was copied from, nesting included. Put the assets the screen references under `screens/` at the paths it asks for, or inline them as data URIs. Anything resolving outside `screens/` is refused.
 
 Before handing over the URL, look at the rendered screen — a screenshot where the platform has one, otherwise measure the laid-out result in the DOM. A 200 on every asset is not that check: an image that loads correctly at the wrong size passes it, as does a script that leaves the page inert. Check each variant at rest, not just the page — one bug in shared scaffolding reads as several bad designs. Drive an interaction only when its behavior is invisible at rest, which is also the case where telling them to try something you have not tried is a claim you made up. Measurement lies by default — computed styles read mid-transition, scroll events coalesce — so read after things settle, and suspect the instrument before you conclude the page is broken. You are done when they could judge the idea, not when the code is correct. If you have no way to see the rendered result, say so when you hand over the URL rather than implying it was checked.
 
@@ -111,7 +111,20 @@ The fallback root takes the same shape under `/tmp/compound-engineering-<uid>/ce
 
 ## Handoff
 
-Pick the browser that will load the prototype — one surface. Hand it the origin it will actually request. The helper prints `http://localhost:<port>`. Use that when that browser is on this machine. When it is not, start with `--host 0.0.0.0` and hand the explorer the helper's returned URL with only the host rewritten to one they can reach. Rewrite only the host. Do not also hand localhost. Visiting that origin sets the session cookie. Wait talks the bind address with the file token. Do not print the token. Binding every interface serves the run directory to anything that can reach the port; annotation and wait routes stay token-gated — do it only on a network the user trusts, and say so when you hand over the URL.
+Pick the browser that will load the prototype — one surface. For an annotated preview, use the returned `authorize_url`, not the public `url`; the browser must establish its session cookie before it can annotate. A plain preview still uses `url`.
+
+To select an initial screen or application state, set the authorization link's optional `next` query parameter to the complete origin-root-relative destination, including its query and fragment. Encode it as one parameter, for example:
+
+```js
+const handoff = new URL(authorizeUrl); // the helper's returned authorize_url
+handoff.searchParams.set("next", "/nested/checkout.html?variant=b&token=demo#details");
+```
+
+Use `handoff.href` for that browser. Without `next`, the destination is `/`. Absolute URLs, protocol-relative URLs (`//host`), backslashes, and control characters are refused. Keep application query/hash state inside `next`; do not replace the authorization link's `token` or append the application's fragment directly to the bootstrap URL. The authorization page replaces its own URL with the selected same-origin destination, preserving the application's own `token` and other state without copying the bootstrap credential into it.
+
+When that browser is on this machine, keep the returned host. When it is not, start with `--host 0.0.0.0` and rewrite only the handoff URL's host to one the explorer can reach; preserve its authorization path and complete query, including `token` and any encoded `next`. Do not also hand localhost. Wait talks the bind address with the file token.
+
+Treat `authorize_url` as a bearer credential: deliver it privately to the intended explorer or browser, not to public comments, committed capsules, or shared logs. Do not print the token separately. Anyone holding that link can control the annotation session until it ends. Binding every interface exposes screen files to anything that can reach the port; control-route authentication does not make the screens confidential or encrypt HTTP traffic. Do it only on a network the user trusts, and say so when you hand over the URL.
 
 Keep the server alive the way this host actually does. Detached `start` is the default. If this host reaps detached processes or the URL dies after the tool call, append `--foreground` through its long-running terminal.
